+++ src/Winium.Cruciatus/Core/DisplayOrientation.cs
@@ -1,0 +1,13 @@
+namespace Winium.Cruciatus.Core
+{
+    /// <summary>
+    /// Represents the orientation of the screen
+    /// </summary>
+    public enum DisplayOrientation
+    {
+        LANDSCAPE = 0,
+        LANDSCAPE_FLIPPED = 2,
+        PORTRAIT = 1,
+        PORTRAIT_FLIPPED = 3
+    }
+}