+++ src/Winium.Cruciatus/Helpers/AutomationElementHelper.cs
@@ -20,24 +20,12 @@ namespace Winium.Cruciatus.Helpers
         #region Methods
 
         internal static IEnumerable<AutomationElement> FindAll(
-            AutomationElement parent, 
-            TreeScope scope, 
-            Condition condition, 
+            AutomationElement parent,
+            TreeScope scope,
+            Condition condition,
             int timeout)
         {
-            var dtn = DateTime.Now.AddMilliseconds(timeout);
-
-            // ReSharper disable once LoopVariableIsNeverChangedInsideLoop
-            while (DateTime.Now <= dtn)
-            {
-                var elements = parent.FindAll(scope, condition);
-                if (elements.Count > 0)
-                {
-                    return elements.Cast<AutomationElement>();
-                }
-            }
-
-            return Enumerable.Empty<AutomationElement>();
+            return TreeWalkerFindHelper.FindAll(parent, scope, condition, timeout);
         }
 
         internal static IEnumerable<AutomationElement> FindAll(AutomationElement parent, string xpath, int timeout)
@@ -73,24 +61,12 @@ namespace Winium.Cruciatus.Helpers
         }
 
         internal static AutomationElement FindFirst(
-            AutomationElement parent, 
-            TreeScope scope, 
-            Condition condition, 
+            AutomationElement parent,
+            TreeScope scope,
+            Condition condition,
             int timeout)
         {
-            var dtn = DateTime.Now.AddMilliseconds(timeout);
-
-            // ReSharper disable once LoopVariableIsNeverChangedInsideLoop
-            while (DateTime.Now <= dtn)
-            {
-                var element = parent.FindFirst(scope, condition);
-                if (element != null)
-                {
-                    return element;
-                }
-            }
-
-            return null;
+            return TreeWalkerFindHelper.FindFirst(parent, scope, condition, timeout);
         }
 
         internal static bool TryGetBoundingRectangleCenter(AutomationElement element, out Point point)