+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 <!--## Unreleased-->
 
-## v2.9.0
+## v2.10.0
+
+- Add `RotationManager` for get and set display orientation (thanks to @jonstoneman)
+
+
+## v2.9.1
+
+- Fix XPath attribute value if this ControlType
+
+
+## v2.9.0
 
 - Add 'SetFocus' method to 'CruciatusElement'
 
@@ -16,5 +26,7 @@
 
 - Set default search timeout to 10 seconds
 - Simplify DataGrid element instantiation
+
+